requirements.txt
streamlit
openai
faiss-cpu
numpy
pandas
scikit-learn
